ci: add CI + Windows release workflows, commit real app icons
- CI: typecheck, frontend build, rustfmt, clippy -D warnings, cargo test (push + PRs)
- Release: auto-build NSIS installer on windows-latest and publish a nightly GitHub Release on every push to main
- Generate real app icons (indigo gradient + TA monogram) from app-icon.png
- Fix clippy unnecessary_cast warning, apply rustfmt
- Un-ignore src-tauri/icons/ (required for bundling)

+++ src-tauri/src/db/migrations.rs
@@ -104,7 +104,7 @@ pub fn run_pending(conn: &Connection) -> Result<(), String> {
     let all = get_migrations();
 
     for migration in &all {
-        if applied.contains(&(migration.version as i64)) {
+        if applied.contains(&migration.version) {
             continue; // already applied
         }
